# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/retailers": {
      "get": {
        "operationId": "RetailersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all retailers.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RetailerResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all retailers",
        "tags": [
          "retailers"
        ]
      },
      "post": {
        "operationId": "RetailersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Data to create a new retailer.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRetailerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The retailer has been created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetailerCreatedResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data — the slug or the tax ID is already taken."
          }
        },
        "summary": "Create a new retailer",
        "tags": [
          "retailers"
        ]
      }
    },
    "/retailers/{id}": {
      "get": {
        "operationId": "RetailersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Retailer ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retailer found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetailerResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Retailer not found."
          }
        },
        "summary": "Get a single retailer",
        "tags": [
          "retailers"
        ]
      },
      "patch": {
        "operationId": "RetailersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID of the retailer to update",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "New data for the retailer.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRetailerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Retailer updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetailerCreatedResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data — the slug or the tax ID is already taken."
          },
          "404": {
            "description": "Retailer not found."
          }
        },
        "summary": "Update a retailer",
        "tags": [
          "retailers"
        ]
      },
      "delete": {
        "operationId": "RetailersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Retailer ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retailer deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Retailer not found."
          }
        },
        "summary": "Delete a retailer by its ID",
        "tags": [
          "retailers"
        ]
      }
    },
...
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all users.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all users",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Data to create a new user.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The user has been created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCreatedResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data — duplicated email, or profile link does not match the role."
          }
        },
        "summary": "Create a new user",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found."
          }
        },
        "summary": "Get a single user",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID of the user to update",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "New data for the user.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCreatedResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data — duplicated email, or profile link does not match the role."
          },
          "404": {
            "description": "User not found."
          }
        },
        "summary": "Update a user",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found."
          }
        },
        "summary": "Delete a user by its ID",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Spira API Documentation",
    "description": "Spira API",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "FoodCategory": {
        "type": "string",
        "enum": [
          "PRODUCE",
          "BAKERY",
          "DAIRY",
          "MEAT",
          "PREPARED",
          "DRY_GOODS",
          "FROZEN"
        ],
        "description": "Food categories typically donated."
      },
...
      "RetailerResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique retailer ID.",
            "format": "uuid"
          },
          "legalName": {
            "type": "string",
            "description": "Registered company name.",
            "example": "Supermercados Real S.A."
          },
          "tradeName": {
            "type": "object",
            "description": "Public brand.",
            "nullable": true,
            "example": "Real"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly identifier.",
            "example": "supermercados-real"
          },
          "taxId": {
            "type": "string",
            "description": "Government tax identifier.",
            "example": "80012345-6"
          },
          "businessType": {
            "description": "Kind of business.",
            "example": "SUPERMARKET",
            "allOf": [
              {
                "$ref": "#/components/schemas/BusinessType"
              }
            ]
          },
          "description": {
            "type": "object",
            "description": "Company description.",
            "nullable": true
          },
          "website": {
            "type": "object",
            "description": "Public website.",
            "nullable": true
          },
          "logoUrl": {
            "type": "object",
            "description": "Absolute URL of the logo.",
            "nullable": true
          },
          "foodCategories": {
            "type": "array",
            "description": "Food categories typically donated.",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/FoodCategory"
            }
          },
          "donationFrequency": {
            "description": "How often surplus is expected.",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/DonationFrequency"
              }
            ]
          },
          "requiresRecipientTransport": {
            "type": "boolean",
            "description": "Whether the recipient must collect the donation.",
            "example": true
          },
          "minPickupNoticeHours": {
            "type": "number",
            "description": "Lead time needed before a collection, in hours.",
            "nullable": true
          },
          "handlingInstructions": {
            "type": "object",
            "description": "Notes for whoever collects.",
            "nullable": true
          },
          "status": {
            "description": "Lifecycle status.",
            "example": "PENDING_VERIFICATION",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProfileStatus"
              }
            ]
          },
          "isVerified": {
            "type": "boolean",
            "description": "Whether an admin has verified this retailer.",
            "example": false
          },
          "verifiedAt": {
            "type": "string",
            "description": "When the retailer was verified (ISO 8601).",
            "format": "date-time",
            "nullable": true
          },
          "verifiedById": {
            "type": "object",
            "description": "ID of the admin who verified this retailer.",
            "format": "uuid",
            "nullable": true
          },
          "foodSafetyLicenseNumber": {
            "type": "object",
            "description": "Food-safety licence number.",
            "nullable": true
          },
          "foodSafetyLicenseExpiresAt": {
            "type": "string",
            "description": "Licence expiry date.",
            "format": "date",
            "nullable": true
          },
          "termsAcceptedAt": {
            "type": "string",
            "description": "When the platform terms were accepted (ISO 8601).",
            "format": "date-time",
            "nullable": true
          },
          "termsVersion": {
            "type": "object",
            "description": "Version of the terms accepted.",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "When the retailer was created (ISO 8601).",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "When the retailer was last modified (ISO 8601).",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "legalName",
          "slug",
          "taxId",
          "businessType",
          "requiresRecipientTransport",
          "status",
          "isVerified",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateRetailerDto": {
        "type": "object",
        "properties": {
          "legalName": {
            "type": "string",
            "description": "Registered company name.",
            "example": "Supermercados Real S.A.",
            "minLength": 2,
            "maxLength": 200
          },
          "tradeName": {
            "type": "string",
            "description": "Public brand, when it differs from the legal name.",
            "example": "Real",
            "maxLength": 200
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly identifier. Must be unique.",
            "example": "supermercados-real",
            "maxLength": 120
          },
          "taxId": {
            "type": "string",
            "description": "Government tax identifier (RUC/NIT/CIF/EIN). Must be unique.",
            "example": "80012345-6",
            "maxLength": 40
          },
          "businessType": {
            "description": "Kind of business the retailer runs.",
            "example": "SUPERMARKET",
            "allOf": [
              {
                "$ref": "#/components/schemas/BusinessType"
              }
            ]
          },
          "description": {
            "type": "string",
            "description": "Free-text presentation of the company."
          },
          "website": {
            "type": "string",
            "description": "Public website.",
            "example": "https://real.com.py",
            "maxLength": 255
          },
          "logoUrl": {
            "type": "string",
            "description": "Absolute URL of the logo.",
            "example": "https://cdn.spira.app/logos/supermercados-real.png",
            "maxLength": 255
          },
          "foodCategories": {
            "type": "array",
            "description": "Food categories this retailer typically donates.",
            "example": [
              "PRODUCE",
              "BAKERY"
            ],
            "items": {
              "$ref": "#/components/schemas/FoodCategory"
            }
          },
          "donationFrequency": {
            "description": "How often surplus is expected.",
            "example": "DAILY",
            "allOf": [
              {
                "$ref": "#/components/schemas/DonationFrequency"
              }
            ]
          },
          "requiresRecipientTransport": {
            "type": "boolean",
            "description": "Whether the recipient must collect the donation itself.",
            "default": true,
            "example": true
          },
          "minPickupNoticeHours": {
            "type": "number",
            "description": "Lead time needed before a collection, in hours.",
            "example": 4,
            "minimum": 0,
            "maximum": 720
          },
          "handlingInstructions": {
            "type": "string",
            "description": "Practical notes for whoever collects the donation.",
            "example": "Enter through the loading dock on Calle Palma and ask for the shift manager."
          },
          "foodSafetyLicenseNumber": {
            "type": "string",
            "description": "Food-safety licence number.",
            "maxLength": 80
          },
          "foodSafetyLicenseExpiresAt": {
            "type": "string",
            "description": "Licence expiry date (calendar date, no time).",
            "format": "date",
            "example": "2027-03-31"
          },
          "termsAcceptedAt": {
            "type": "string",
            "description": "When the retailer accepted the platform terms.",
            "format": "date-time",
            "example": "2026-09-16T14:32:05.123Z"
          },
          "termsVersion": {
            "type": "string",
            "description": "Version of the terms that was accepted.",
            "example": "2026-01",
            "maxLength": 20
          }
        },
        "required": [
          "legalName",
          "slug",
          "taxId",
          "businessType"
        ]
      },
      "RetailerCreatedResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique retailer ID.",
            "format": "uuid"
          },
          "legalName": {
            "type": "string",
            "description": "Registered company name.",
            "example": "Supermercados Real S.A."
          },
          "tradeName": {
            "type": "object",
            "description": "Public brand.",
            "nullable": true,
            "example": "Real"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly identifier.",
            "example": "supermercados-real"
          },
          "taxId": {
            "type": "string",
            "description": "Government tax identifier.",
            "example": "80012345-6"
          },
          "businessType": {
            "description": "Kind of business.",
            "example": "SUPERMARKET",
            "allOf": [
              {
                "$ref": "#/components/schemas/BusinessType"
              }
            ]
          },
          "description": {
            "type": "object",
            "description": "Company description.",
            "nullable": true
          },
          "website": {
            "type": "object",
            "description": "Public website.",
            "nullable": true
          },
          "logoUrl": {
            "type": "object",
            "description": "Absolute URL of the logo.",
            "nullable": true
          },
          "foodCategories": {
            "type": "array",
            "description": "Food categories typically donated.",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/FoodCategory"
            }
          },
          "donationFrequency": {
            "description": "How often surplus is expected.",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/DonationFrequency"
              }
            ]
          },
          "requiresRecipientTransport": {
            "type": "boolean",
            "description": "Whether the recipient must collect the donation.",
            "example": true
          },
          "minPickupNoticeHours": {
            "type": "number",
            "description": "Lead time needed before a collection, in hours.",
            "nullable": true
          },
          "handlingInstructions": {
            "type": "object",
            "description": "Notes for whoever collects.",
            "nullable": true
          },
          "status": {
            "description": "Lifecycle status.",
            "example": "PENDING_VERIFICATION",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProfileStatus"
              }
            ]
          },
          "isVerified": {
            "type": "boolean",
            "description": "Whether an admin has verified this retailer.",
            "example": false
          },
          "verifiedAt": {
            "type": "string",
            "description": "When the retailer was verified (ISO 8601).",
            "format": "date-time",
            "nullable": true
          },
          "verifiedById": {
            "type": "object",
            "description": "ID of the admin who verified this retailer.",
            "format": "uuid",
            "nullable": true
          },
          "foodSafetyLicenseNumber": {
            "type": "object",
            "description": "Food-safety licence number.",
            "nullable": true
          },
          "foodSafetyLicenseExpiresAt": {
            "type": "string",
            "description": "Licence expiry date.",
            "format": "date",
            "nullable": true
          },
          "termsAcceptedAt": {
            "type": "string",
            "description": "When the platform terms were accepted (ISO 8601).",
            "format": "date-time",
            "nullable": true
          },
          "termsVersion": {
            "type": "object",
            "description": "Version of the terms accepted.",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "When the retailer was created (ISO 8601).",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "When the retailer was last modified (ISO 8601).",
            "format": "date-time"
          },
          "message": {
            "type": "string",
            "description": "Human-readable message describing the outcome.",
            "example": "Retailer created successfully."
          }
        },
        "required": [
          "id",
          "legalName",
          "slug",
          "taxId",
          "businessType",
          "requiresRecipientTransport",
          "status",
          "isVerified",
          "createdAt",
          "updatedAt",
          "message"
        ]
      },
...
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Human-readable message describing the outcome.",
            "example": "Request completed successfully"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique user ID.",
            "format": "uuid",
            "example": "3f2c1b8e-9a4d-4c7f-8b1e-2d6a5c9f0e11"
          },
          "email": {
            "type": "string",
            "description": "Login email.",
            "example": "maria.gonzalez@real.com.py"
          },
          "role": {
            "description": "Authorization role.",
            "example": "RETAILER",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "retailerId": {
            "type": "object",
            "description": "Linked retailer, when `role` is `RETAILER`.",
            "format": "uuid",
            "nullable": true,
            "example": null
          },
          "recipientId": {
            "type": "object",
            "description": "Linked recipient, when `role` is `RECIPIENT`.",
            "format": "uuid",
            "nullable": true,
            "example": null
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the account may log in.",
            "example": true
          },
          "isEmailVerified": {
            "type": "boolean",
            "description": "Whether the email has been confirmed.",
            "example": false
          },
          "emailVerifiedAt": {
            "type": "string",
            "description": "When the email was confirmed (ISO 8601).",
            "format": "date-time",
            "nullable": true,
            "example": null
          },
          "lastLoginAt": {
            "type": "string",
            "description": "Last successful login (ISO 8601).",
            "format": "date-time",
            "nullable": true,
            "example": null
          },
          "createdAt": {
            "type": "string",
            "description": "When the account was created (ISO 8601).",
            "format": "date-time",
            "example": "2026-09-16T14:32:05.123Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "When the account was last modified (ISO 8601).",
            "format": "date-time",
            "example": "2026-09-16T14:32:05.123Z"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "isActive",
          "isEmailVerified",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}